Module 13 start

+++ Outside_Module/tempCodeRunnerFile.c
@@ -4,7 +4,7 @@
    long long int toy=0;
    scanf("%lld %lld %lld",&e,&m,&b);
    long long int sum = e+b;
-   for (long long int i = 0; i <= sum; i++)
+   for (int i = 0; i <= sum; i++)
     {
         if (e>=1 && m>=1 && b>=1)
         {
